asr_test_top.sh: add mode and -n argument parsing
Allows running specific test modes e.g. `./asr_test_top.sh radev2 -n 100`
instead of always running all tests. Removes hardcoded exit 0 that was
blocking radev2/ssb/rade sweeps from running.

+++ asr_test_top.sh
@@ -5,6 +5,14 @@
 set -x
 results_file=241221_asr
 n=500
+mode=${1:-all}
+shift
+while [[ $# -gt 0 ]]; do
+    case $1 in
+        -n) n=$2; shift 2;;
+        *) shift;;
+    esac
+done
 
 function ssb {
     local results_file=$1
@@ -50,28 +58,33 @@ function freedv_700D {
     mv tmp.txt ${results_file}
 }
 
-freedv_700D ${results_file}_awgn_700D.txt  "-100 -30 -26 -23 -20 -17 -15 -13"
-freedv_700D ${results_file}_mpp_700D.txt   "-100 -39 -36 -33 -30 -27" "--g_file g_mpp.f32"
-#freedv_700D ${results_file}_awgn_700D.txt  "-100 -38 -35 -32 -29 -26 -23 -20 -17"
-#freedv_700D ${results_file}_mpp_700D.txt   "-100 -44 -39 -36 -33 -30 -27" "--g_file g_mpp.f32"
-exit 0
+if [ "$mode" == "700D" ] || [ "$mode" == "all" ]; then
+    freedv_700D ${results_file}_awgn_700D.txt  "-100 -30 -26 -23 -20 -17 -15 -13"
+    freedv_700D ${results_file}_mpp_700D.txt   "-100 -39 -36 -33 -30 -27" "--g_file g_mpp.f32"
+fi
 
-# run the controls
-controls_file=${results_file}_controls.txt
-rm -f ${controls_file}
-./asr_test.sh clean -n $n --results ${controls_file}
-./asr_test.sh fargan -n $n --results ${controls_file}
-./asr_test.sh 4kHz -n $n --results ${controls_file}
-./asr_test.sh ssb -n $n --results ${controls_file}
-./asr_test.sh rade -n $n --results ${controls_file}
-# strip off all but last column for Octave plotting
-cat ${controls_file} | awk '{print $NF}' > ${results_file}_c.txt
+if [ "$mode" == "controls" ] || [ "$mode" == "all" ]; then
+    controls_file=${results_file}_controls.txt
+    rm -f ${controls_file}
+    ./asr_test.sh clean -n $n --results ${controls_file}
+    ./asr_test.sh fargan -n $n --results ${controls_file}
+    ./asr_test.sh 4kHz -n $n --results ${controls_file}
+    # strip off all but last column for Octave plotting
+    cat ${controls_file} | awk '{print $NF}' > ${results_file}_c.txt
+fi
 
+if [ "$mode" == "ssb" ] || [ "$mode" == "all" ]; then
+    ssb    ${results_file}_awgn_ssb.txt    "-100 -38 -35 -32 -29 -26 -23 -20 -17"
+    ssb    ${results_file}_mpp_ssb.txt    "-100 -44 -39 -36 -33 -30 -27" "--g_file g_mpp.f32"
+fi
 
-ssb    ${results_file}_awgn_ssb.txt    "-100 -38 -35 -32 -29 -26 -23 -20 -17"
-rade   ${results_file}_awgn_rade.txt   "100 15 10 5 2.5 0 -2.5"
-radev2 ${results_file}_awgn_radev2.txt "100 15 10 5 2.5 0 -2.5"
-ssb    ${results_file}_mpp_ssb.txt    "-100 -44 -39 -36 -33 -30 -27" "--g_file g_mpp.f32"
-rade   ${results_file}_mpp_rade.txt   "100 15 10 5 2.5 0" "--g_file g_mpp.f32"
-radev2 ${results_file}_mpp_radev2.txt "100 15 10 5 2.5 0" "--g_file g_mpp.f32"
+if [ "$mode" == "rade" ] || [ "$mode" == "all" ]; then
+    rade   ${results_file}_awgn_rade.txt   "100 15 10 5 2.5 0 -2.5"
+    rade   ${results_file}_mpp_rade.txt   "100 15 10 5 2.5 0" "--g_file g_mpp.f32"
+fi
+
+if [ "$mode" == "radev2" ] || [ "$mode" == "all" ]; then
+    radev2 ${results_file}_awgn_radev2.txt "100 15 10 5 2.5 0 -2.5"
+    radev2 ${results_file}_mpp_radev2.txt  "100 15 10 5 2.5 0" "--g_file g_mpp.f32"
+fi
